console: Rename file and merge tti parts
