+++ pkg/webui/console/containers/gateway-onboarding-form/qr-scan-section/fleets-scan.js
