+++ pkg/webui/console/containers/gateway-onboarding-form/qr-scan-section/index.js
@@ -31,7 +31,7 @@ import sharedMessages from '@ttn-lw/lib/shared-messages'
 import { parseGatewayQRCode } from '@console/store/actions/qr-code-generator'
 import { getGatewayClaimInfoByEui } from '@console/store/actions/gateways'
 
-import FleetsScan from './fleets-scan.tti'
+import FleetsScan from './fleets-scan'
 
 const m = defineMessages({
   hasGatewayQR:
@@ -70,7 +70,6 @@ const GatewayQRScanSection = () => {
       },
       authenticated_identifiers: {
         gateway_eui: gateway.gateway_eui,
-        // TTI only.
         authentication_code: gateway._fleet_owner_token
           ? btoa(gateway._fleet_owner_token)
           : gateway.owner_token
@@ -80,7 +79,6 @@ const GatewayQRScanSection = () => {
       _gtw_owner_token: gateway.owner_token ?? '',
       _fleet_owner_token: gateway._fleet_owner_token ?? '',
       _isFleet: Boolean(gateway._fleet_owner_token),
-      // End TTI.
     }))
 
     setQrData({ ...qrData, approved: true })
@@ -96,10 +94,8 @@ const GatewayQRScanSection = () => {
         // Get gateway from QR code
         const gateway = await dispatch(attachPromise(parseGatewayQRCode(qrCode)))
         const gatewayEui = gateway.gateway_eui
-        // TTI only.
         const { is_managed } = await dispatch(attachPromise(getGatewayClaimInfoByEui(gatewayEui)))
         gateway.is_managed = is_managed
-        // End TTI.
 
         const sheetData = [
           {
@@ -170,9 +166,7 @@ const GatewayQRScanSection = () => {
             onCancel={handleQRCodeCancel}
             onRead={handleQRCodeRead}
             qrData={qrData}
-            // TTI only.
             modalDataChildren={<FleetsScan qrData={qrData} setQrData={setQrData} />}
-            // End TTI.
           />
         )}
       </ButtonGroup>
